refactor: Make new UI the main landing page
Major changes:
- Removed old static/index.html and static/index-modular.html
- Moved new UI from /ui to root path (/)
- Consolidated file structure:
  - static/index.html is now the new UI
  - static/js/modules/ contains all UI modules
  - static/css/app.css contains UI styles
  - static/js/globals.js contains global functions

Technical updates:
- Root endpoint (/) now redirects to /static/index.html
- Removed /ui endpoint as it's no longer needed
- Updated CLAUDE.md with new access URLs

Access URLs:
- Main UI: http://localhost:8000
- API docs: http://localhost:8000/docs
- Visualizations: http://localhost:8000/static/visualizations.html

All functionality remains intact with cleaner URL structure.
