+++ static/js/globals.js
