+++ static/js/modules/configurationManager.js
