+++ static/js/modules/templateManager.js
