--- static/js/modules/utils.js
@@ -1,201 +1,0 @@
-/**
- * Utility Functions Module
- * Common utility functions for the application
- */
-
-/**
- * Debounce function to limit the rate of function calls
- * @param {Function} func - Function to debounce
- * @param {number} wait - Wait time in milliseconds
- * @param {boolean} immediate - Trigger on leading edge instead of trailing
- * @returns {Function} Debounced function
- */
-export function debounce(func, wait, immediate = false) {
-    let timeout;
-    
-    return function executedFunction(...args) {
-        const context = this;
-        
-        const later = () => {
-            timeout = null;
-            if (!immediate) func.apply(context, args);
-        };
-        
-        const callNow = immediate && !timeout;
-        clearTimeout(timeout);
-        timeout = setTimeout(later, wait);
-        
-        if (callNow) func.apply(context, args);
-    };
-}
-
-/**
- * Throttle function to limit function calls to once per specified time
- * @param {Function} func - Function to throttle
- * @param {number} limit - Time limit in milliseconds
- * @returns {Function} Throttled function
- */
-export function throttle(func, limit) {
-    let inThrottle;
-    
-    return function(...args) {
-        const context = this;
-        
-        if (!inThrottle) {
-            func.apply(context, args);
-            inThrottle = true;
-            setTimeout(() => inThrottle = false, limit);
-        }
-    };
-}
-
-/**
- * Retry function with exponential backoff
- * @param {Function} fn - Async function to retry
- * @param {number} maxRetries - Maximum number of retries
- * @param {number} baseDelay - Base delay in milliseconds
- * @param {number} maxDelay - Maximum delay in milliseconds
- * @returns {Promise} Result of the function
- */
-export async function retryWithBackoff(fn, maxRetries = 3, baseDelay = 1000, maxDelay = 30000) {
-    let retries = 0;
-    
-    while (retries < maxRetries) {
-        try {
-            return await fn();
-        } catch (error) {
-            retries++;
-            
-            if (retries === maxRetries) {
-                throw error;
-            }
-            
-            // Calculate exponential backoff with jitter
-            const delay = Math.min(
-                baseDelay * Math.pow(2, retries - 1) + Math.random() * 1000,
-                maxDelay
-            );
-            
-            console.log(`Retry ${retries}/${maxRetries} after ${delay}ms`);
-            await sleep(delay);
-        }
-    }
-}
-
-/**
- * Sleep function for delays
- * @param {number} ms - Milliseconds to sleep
- * @returns {Promise} Promise that resolves after the delay
- */
-export function sleep(ms) {
-    return new Promise(resolve => setTimeout(resolve, ms));
-}
-
-/**
- * Deep clone an object
- * @param {*} obj - Object to clone
- * @returns {*} Cloned object
- */
-export function deepClone(obj) {
-    if (obj === null || typeof obj !== 'object') return obj;
-    if (obj instanceof Date) return new Date(obj.getTime());
-    if (obj instanceof Array) return obj.map(item => deepClone(item));
-    if (obj instanceof Object) {
-        const clonedObj = {};
-        for (const key in obj) {
-            if (obj.hasOwnProperty(key)) {
-                clonedObj[key] = deepClone(obj[key]);
-            }
-        }
-        return clonedObj;
-    }
-}
-
-/**
- * Format bytes to human readable string
- * @param {number} bytes - Number of bytes
- * @param {number} decimals - Number of decimal places
- * @returns {string} Formatted string
- */
-export function formatBytes(bytes, decimals = 2) {
-    if (bytes === 0) return '0 Bytes';
-    
-    const k = 1024;
-    const dm = decimals < 0 ? 0 : decimals;
-    const sizes = ['Bytes', 'KB', 'MB', 'GB', 'TB'];
-    
-    const i = Math.floor(Math.log(bytes) / Math.log(k));
-    
-    return parseFloat((bytes / Math.pow(k, i)).toFixed(dm)) + ' ' + sizes[i];
-}
-
-/**
- * Generate a unique ID
- * @returns {string} Unique ID
- */
-export function generateId() {
-    return `${Date.now()}-${Math.random().toString(36).substr(2, 9)}`;
-}
-
-/**
- * Check if value is empty
- * @param {*} value - Value to check
- * @returns {boolean} True if empty
- */
-export function isEmpty(value) {
-    if (value == null) return true;
-    if (typeof value === 'string') return value.trim().length === 0;
-    if (Array.isArray(value)) return value.length === 0;
-    if (typeof value === 'object') return Object.keys(value).length === 0;
-    return false;
-}
-
-/**
- * Memoize function results
- * @param {Function} fn - Function to memoize
- * @returns {Function} Memoized function
- */
-export function memoize(fn) {
-    const cache = new Map();
-    
-    return function(...args) {
-        const key = JSON.stringify(args);
-        
-        if (cache.has(key)) {
-            return cache.get(key);
-        }
-        
-        const result = fn.apply(this, args);
-        cache.set(key, result);
-        
-        return result;
-    };
-}
-
-/**
- * Parse query string to object
- * @param {string} queryString - Query string to parse
- * @returns {Object} Parsed query parameters
- */
-export function parseQueryString(queryString) {
-    const params = new URLSearchParams(queryString);
-    const result = {};
-    
-    for (const [key, value] of params) {
-        result[key] = value;
-    }
-    
-    return result;
-}
-
-/**
- * Create query string from object
- * @param {Object} params - Parameters object
- * @returns {string} Query string
- */
-export function createQueryString(params) {
-    return Object.entries(params)
-        .filter(([_, value]) => value != null)
-        .map(([key, value]) => `${encodeURIComponent(key)}=${encodeURIComponent(value)}`)
-        .join('&');
-}
