--- static/new-ui/index.html
@@ -1,155 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Military Patient Generator</title>
-    
-    <!-- Bootstrap 5 CSS -->
-    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/css/bootstrap.min.css" rel="stylesheet">
-    <!-- Bootstrap Icons -->
-    <link href="https://cdn.jsdelivr.net/npm/bootstrap-icons@1.11.0/font/bootstrap-icons.css" rel="stylesheet">
-    <!-- Custom CSS -->
-    <link href="css/app.css" rel="stylesheet">
-</head>
-<body>
-    <!-- Header -->
-    <nav class="navbar navbar-dark bg-primary">
-        <div class="container-fluid">
-            <span class="navbar-brand mb-0 h1">
-                <i class="bi bi-hospital me-2"></i>Military Patient Generator
-            </span>
-            <div class="d-flex align-items-center">
-                <!-- Quick Start button hidden for future development
-                <button class="btn btn-outline-light me-2" id="quickStartBtn">
-                    <i class="bi bi-lightning-fill me-1"></i>Quick Start
-                </button>
-                -->
-                <button class="btn btn-outline-light" id="helpBtn">
-                    <i class="bi bi-question-circle"></i>
-                </button>
-            </div>
-        </div>
-    </nav>
-
-    <!-- Main Content -->
-    <div class="container-fluid mt-3">
-        <div class="row">
-            <!-- Configuration Panel -->
-            <div class="col-lg-8">
-                <div class="card shadow-sm">
-                    <div class="card-header bg-light">
-                        <h5 class="mb-0">
-                            <i class="bi bi-gear me-2"></i>Configuration
-                        </h5>
-                    </div>
-                    <div class="card-body">
-                        <!-- Basic Settings -->
-                        <div id="basicSettings" class="config-section"></div>
-                        
-                        <!-- Medical Facility Defaults -->
-                        <div id="medicalFacilityDefaults" class="config-section mt-4"></div>
-                        
-                        <!-- Front Configuration -->
-                        <div id="frontConfiguration" class="config-section mt-4"></div>
-                        
-                        <!-- Injury Distribution -->
-                        <div id="injuryDistribution" class="config-section mt-4"></div>
-                        
-                        <!-- Output Configuration -->
-                        <div id="outputConfiguration" class="config-section mt-4"></div>
-                        
-                        <!-- Template System -->
-                        <div id="templateSystem" class="config-section mt-4"></div>
-                        
-                        <!-- Action Buttons -->
-                        <div class="d-flex justify-content-between mt-4">
-                            <button class="btn btn-secondary" id="resetAllBtn">
-                                <i class="bi bi-arrow-clockwise me-1"></i>Reset All
-                            </button>
-                            <button class="btn btn-primary btn-lg" id="generateBtn">
-                                <i class="bi bi-play-fill me-1"></i>Generate Patients
-                            </button>
-                        </div>
-                    </div>
-                </div>
-            </div>
-            
-            <!-- Jobs Panel -->
-            <div class="col-lg-4">
-                <div class="card shadow-sm">
-                    <div class="card-header bg-light">
-                        <h5 class="mb-0">
-                            <i class="bi bi-clipboard-check me-2"></i>Jobs
-                        </h5>
-                    </div>
-                    <div class="card-body" id="jobsPanel">
-                        <!-- Jobs will be dynamically loaded here -->
-                        <div class="text-center text-muted py-5">
-                            <i class="bi bi-arrow-repeat fs-1"></i>
-                            <p class="mt-2">No jobs yet.<br>Configure and start a generation job.</p>
-                        </div>
-                    </div>
-                </div>
-            </div>
-        </div>
-    </div>
-
-    <!-- Template Selection Modal -->
-    <div class="modal fade" id="templateModal" tabindex="-1">
-        <div class="modal-dialog modal-lg">
-            <div class="modal-content">
-                <div class="modal-header">
-                    <h5 class="modal-title">Configuration Templates</h5>
-                    <button type="button" class="btn-close" data-bs-dismiss="modal"></button>
-                </div>
-                <div class="modal-body" id="templateList">
-                    <!-- Template cards will be dynamically loaded here -->
-                </div>
-                <div class="modal-footer">
-                    <button type="button" class="btn btn-secondary" data-bs-dismiss="modal">Continue with Blank</button>
-                    <button type="button" class="btn btn-secondary" data-bs-dismiss="modal">Close</button>
-                </div>
-            </div>
-        </div>
-    </div>
-
-    <!-- Progress Modal -->
-    <div class="modal fade" id="progressModal" tabindex="-1">
-        <div class="modal-dialog">
-            <div class="modal-content">
-                <div class="modal-header">
-                    <h5 class="modal-title">Generating Patients</h5>
-                </div>
-                <div class="modal-body">
-                    <div class="mb-3">
-                        <div class="progress" style="height: 30px;">
-                            <div class="progress-bar progress-bar-striped progress-bar-animated" 
-                                 id="progressBar" 
-                                 role="progressbar" 
-                                 style="width: 0%">0%</div>
-                        </div>
-                    </div>
-                    <div id="progressDetails">
-                        <p class="mb-1"><strong>Status:</strong> <span id="progressStatus">Initializing...</span></p>
-                        <p class="mb-1"><strong>Patients:</strong> <span id="progressPatients">0/0</span></p>
-                        <p class="mb-0"><strong>Phase:</strong> <span id="progressPhase">Starting generation...</span></p>
-                    </div>
-                </div>
-                <div class="modal-footer">
-                    <button type="button" class="btn btn-secondary" id="cancelJobBtn">Cancel</button>
-                </div>
-            </div>
-        </div>
-    </div>
-
-    <!-- Bootstrap JS Bundle -->
-    <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/js/bootstrap.bundle.min.js"></script>
-    
-    <!-- Global functions -->
-    <script src="js/globals.js"></script>
-    
-    <!-- Application Scripts -->
-    <script type="module" src="js/app.js"></script>
-</body>
-</html>
